Add in basic autogeneration of themes
This is working on Auto generate themes #14, it follows the alternative proposed soltuion.

It also clears the theme descriptions outlined in Themes to give weightings to #3 as the themes are now part of output so moved into the respective folder.

These genereated theme might be good but are not very stable.