57-expand-support-to-all-modes-of-transport
Still problems with the output of the engine. It is not very accurate however there are going to be big changes to the prompts as we get into the reasoning issues.


* Add in support for multi modes in the Extract_Wrangle module

* Add better support for downloading specific number of reports per year.

* Theme generatoin for multiple modes. Update to gpt 4 for some tasks

Appears after spot tests to look reasonable.

* fix #52. Update to support reason extraction. Multi modal support checked but not verified.

* Update the readme and discuss the config file

* Update summarizer to output the full summary with reasoing into the report folder.

* File names were reversed for weightings and the full summary.

* Update local viewer to work.

Needed to add in a run function.

* Revert "Update local viewer to work."

This reverts commit 2d8d965205a2e942c46abbd0d38c192624220874.

This was added on the wrong branch by mistake.

* Update weighting collection prompt. This provides some token reason.

The token reason will be expanded upon for in #56.

* Fix summarizer so it should never crash.

* Make parsed output  the same format as template