55-quote-and-reference-checker
Add in quote checker and repairer.

This has been integrated with ThemeGeneration but not Weightings as issue #56 has not yet been completed,

* Update the theme generator to output the themes into yaml.

This makes it easier for the reference checking and manipulation of the themes.

* Finishing for the the night

* Basic reference checker.

This will extract and check all of the references both quotes and citations given in a chunk of text.

Currently the theme generator does nothing to fix broken references.

* Add in basic reparing of references.

This means that the referencesChecking now returns either validated referneces, repaired references or unrepairable references.

ThemeGenerator will need to be updated to e able to handle the updated references

* Add in fixing of the references in the Reference checker.

It will now return the reapired text.

* Reenable the theme generator openAI api request.

* Add support for multiple section location per reference.

* Fix small changes causing crashes

* Saving work, this commit is not complete. Add citation repair.

* Update theme propmt to support better formatted quotes.

* Add in invalid message. Finish the citation repairing.

* Make the citation checking softer

* Remove debug logouts