Update safety issue prompt to make it clear that all exact safety issues should be extracted.
validation set is at 98% with exact to exact compare. The one missing actaully just a sumamry of all safety issues in the intruduction

validation set is 73% match on exact vs inferred.

It still outputs only inferred but I believe this to be due to the finetuning. I will update that will a retrain and some data that includes exact ones.