130-recommendation-extraction-and-safety-issue-linking (#147)
The linking of rcommendations and safety issues is added into the engine and well as support in the viewer.

Most of the development was done inside the recommendation_safety_issue_linking folder.

The final results showed that there is about a 4% variance between runthroughs. This is support there only 5% of the links actually showed up in any of the variance. This means that there are some regulars who are having problems with the linking.

Commit history

* Fix typo

* Prevent infinite loop in the extraction of important text

* Update the content section reader to use gpt 4

* Add basic notebook to create and analyze the links

* Add back in sSI extraction with regex. It is not working perfectly.

* Update and restructure

waiting for response so uploading for safety

* Update to using gpt 4.

Analyzing results.

Waiting for email response.

* Add link upgrading and cleaning up the rest of analysis.

Also do a complete run through of all of the data

* Add in recommendations to the viewer

* Final linking that can be integrated into engine

* fix error missed saving

* INtegrating it into the engine

* Updating viewer to use proper format

* Add visualization of the links

* Add in link upgrading of possible to confirmed when no other links are found.

* Add recommendations safety issue link csv download

* Comparing the engine output to the notebook output.

I am looking at consistency

* Remove the link dataframe from printout

* Simplify the csv download code

* Checking it shows that there is about 4% variance in each run through.

There were only a 57 problematic links when compared to 3 run throughs. This suggests that the variance is bounded and variating around a centre rather than a random walk.

* Adding comparison code