146-investigate-the-viability-of-a-vector-database-to-solve-the-searching-of-reports-problem (#167)
* Tech demo of working vector database

* Testing out RAG.

I have impletmend a test with langchain. It is getting stuck on a few things. There are a few errors but it also failes to retrievs the right report.

* saving

* Add extraction of each section using ReportExtracting

* Upgrade the section extraction. It now is much more reliable and doesn't have to use the LLM nearly as much.

The only large problem is that some of the text files have pages missing meaing that it extracts more sections then needed.

* Add parsing of the reports to sectionss.

* Get the embeddings for both important text and individual sections.

Next step will be to test out what the searching can look like. Then see what a useful way is to get the relevant sections and relevant reports.

* Finish initial investigation into vector databases and rags.

Proper work is going to commence now in invidivual branches.

* Merge branch 'main' into 146-investigate-the-viability-of-a-vector-database-to-solve-the-searching-of-reports-problem

* Old stashed changes