257-allow-pdf-parsing-to-text-to-handle-atsb-and-tsb-reports (#264)
resolve #257
* I have about 90% accurate page number extraction from the ATSB reports.

However I can do a near perfect job if I use the PDF reader to actually extract the page number directly. Then I calibrate it by finding the written page numbers in the report.

* Add in the usage of the pdf pages to sync the regex page numbers.

I need to add testing to this and fix some bugs.

* Add some tests and finish text extraction of ATSB with about a 5-10% error rate.

I will try to add in TSB and see how it goes. Lastly I need to move over taic to this as I have added page numbers in the text extraction stage.

* Add in support for TSB page number extraction

* Completed TAIC into the same framework. Now all three agencies are added.

A final check of the page extraction should be done.

* Final checks and removing printouts

There is over a 97% success rate.
I have spot checked and the validator seems to work as one would expect.

* Update test