265-content-section-extraction (#270)
* Started work on it. Established what needs to be done.

There is a chance that I will need develop some document crawler create an artificial content section.

* Update the website scraping to speed it up

* Fixing TAIC website scraping

* Fix tests failing

The test_form_submit still fails but that is because it is searching the online database I believe.

* Add some content section extraction tests to check for regression with TAIC extraction.

TAIC_r_2002_122 is failing for now. But hopefully with more work we can get it working.

* Add in content page extraction for ATSB

This works well enough. The content pages that arent extracted are from short reports. In this case the whole report can be considered important text.

* Add headers extraction from pdf outline.

This means that the content sections extraction is more or less solved for now.

About 2/3 have a content section or headers with page numbers. This can be feed into the LLM to find the important text. Otherwise the whole report can be treated as important text.

* Make the tests work with changes in output.

* Fix tests

* Integrate the pdf headers into the get_content_section function.

This leads the way for quick development of #266

* FIx up optional parameters