Add important text extraction
* Add in basic important text extraction from content section

The current problem I am facing is the the important section are too long. I need to change the content section reading prompt and system

* Fix the important text extraction.

It is only about 50% bigger than the previous one. I have moved it to gpt-4o which is more expensive so a complete run through will cost about $50. It also uses the gpt4o-mini to do content section cleaning.

* Fixing up tests

* Make important text extraction multithreaded.