Doing-complete-runthrough (#287)
* Update PyPDF to version 5.1. Furthermore fix up TSB reports being excluded  due to parsing errors.

* Fix up gather so that it works properly as a pipeline.

The report downloading and report titles creation works.

There are some problems with mismatches of ids that mean that 20% of ATSB si are unmatched this is not a huge problem.

Lastly recommendation extraction is straight forward with only a few not matched.

* Get safety issue extraction working

Refactor the code so that important text is extracted differently for Recommendations and safety issues. Where it is the content section that gets saved to a file rather than important text.

Thinking about having a report class as it could help passing things around.

* Fix up recommendations pipeline

* Add in section extraction starting printout

* report type assignment working in pipeline

* Finished extract pipeline

* Update recommendation process to work well wit the recommendation safety issue linking.

* It is only TAIC with reply text so not much needs to be changed.

* Fix up recommendation linking and response classification to make it faster.

* Embedding and engine output uploading

* Adding in more metadata to extracted reports.

* Fixup the webapp to support multiple agencies

* Final run through that seems to be working