Adding-other-agencies (#288)
Fixes #286
Fixes #268
Fixes #283
Fixes #254
Fixes #252
Fixes #244

* 253-web-scraping-of-other-agencies (#258)

* Bug fixes

There were some problems when doing complete run through. As well as when only doing a subset of the modes.

* 257-allow-pdf-parsing-to-text-to-handle-atsb-and-tsb-reports (#264)

resolve #257
* I have about 90% accurate page number extraction from the ATSB reports.

However I can do a near perfect job if I use the PDF reader to actually extract the page number directly. Then I calibrate it by finding the written page numbers in the report.

* Add in the usage of the pdf pages to sync the regex page numbers.

I need to add testing to this and fix some bugs.

* Add some tests and finish text extraction of ATSB with about a 5-10% error rate.

I will try to add in TSB and see how it goes. Lastly I need to move over taic to this as I have added page numbers in the text extraction stage.

* Add in support for TSB page number extraction

* Completed TAIC into the same framework. Now all three agencies are added.

A final check of the page extraction should be done.

* Final checks and removing printouts

There is over a 97% success rate.
I have spot checked and the validator seems to work as one would expect.

* Update test

* 265-content-section-extraction (#270)

* Started work on it. Established what needs to be done.

There is a chance that I will need develop some document crawler create an artificial content section.

* Update the website scraping to speed it up

* Fixing TAIC website scraping

* Fix tests failing

The test_form_submit still fails but that is because it is searching the online database I believe.

* Add some content section extraction tests to check for regression with TAIC extraction.

TAIC_r_2002_122 is failing for now. But hopefully with more work we can get it working.

* Add in content page extraction for ATSB

This works well enough. The content pages that arent extracted are from short reports. In this case the whole report can be considered important text.

* Add headers extraction from pdf outline.

This means that the content sections extraction is more or less solved for now.

About 2/3 have a content section or headers with page numbers. This can be feed into the LLM to find the important text. Otherwise the whole report can be treated as important text.

* Make the tests work with changes in output.

* Fix tests

* Integrate the pdf headers into the get_content_section function.

This leads the way for quick development of #266

* FIx up optional parameters

* Add important text extraction

* Add in basic important text extraction from content section

The current problem I am facing is the the important section are too long. I need to change the content section reading prompt and system

* Fix the important text extraction.

It is only about 50% bigger than the previous one. I have moved it to gpt-4o which is more expensive so a complete run through will cost about $50. It also uses the gpt4o-mini to do content section cleaning.

* Fixing up tests

* Make important text extraction multithreaded.

* Adding in paragraph extraction for all reports that dont have section extraction. (#276)

This works well enough with a good enough size. I can come back to it if it seems important at a later date.

* Setup safety issue extraction to mostly work (#278)

* Add in recommednation scraping and extracting (#281)

* Add in tSb web scraping of recommendations

* Add in the agency id extraction.

This means that the report_titles df can be used to convert project id to agency id.

This is needed by the recommendations extraction.

* Add in ATSB recommendation extraction.

This is not working perfectly as there are some problems that more work could fix as well as #279

* Add recommendation extraction into the pipeline.

* Update the taic recommendations as TAIC recommendation scraping has not yet bee implemented.

* Change tAIC recommendations to be a web scrape

* Refactor the recommendation scrapers to be subclasses.

* Restructure the scraping to always ge the metadata.

* Fixing_tests (#282)

* Fix up datagetting now that TAIC recommendations are scraped.

* Add real report ids where they are needed.

* Fix content section extraction

There was problems where the formatter was not being consistent.

This has been fixed with a prompt update.

* Fix some of the report extracting prompts.

* Fix typo in WebScraping.py

* Doing-complete-runthrough (#287)

* Update PyPDF to version 5.1. Furthermore fix up TSB reports being excluded  due to parsing errors.

* Fix up gather so that it works properly as a pipeline.

The report downloading and report titles creation works.

There are some problems with mismatches of ids that mean that 20% of ATSB si are unmatched this is not a huge problem.

Lastly recommendation extraction is straight forward with only a few not matched.

* Get safety issue extraction working

Refactor the code so that important text is extracted differently for Recommendations and safety issues. Where it is the content section that gets saved to a file rather than important text.

Thinking about having a report class as it could help passing things around.

* Fix up recommendations pipeline

* Add in section extraction starting printout

* report type assignment working in pipeline

* Finished extract pipeline

* Update recommendation process to work well wit the recommendation safety issue linking.

* It is only TAIC with reply text so not much needs to be changed.

* Fix up recommendation linking and response classification to make it faster.

* Embedding and engine output uploading

* Adding in more metadata to extracted reports.

* Fixup the webapp to support multiple agencies

* Final run through that seems to be working

* Create analysis of where reports are ending up

* Merge main into the branch, resolve merge conflicts

* Merge branch 'main' into adding-other-agencies

* Fixup tests

* Changing tests and test data

* Fixing up report extraction of content page cleaning and reading.

* Fixing up embedding to make it work with report texts and large texts.

* Some basic evaluation of the searching. It is setup to provide NDCG of the two different embedding of voyage-3 and voyage-large-2-instruct.

The older model does slightly better. For now I will stick with it and then wait for voyage-large-3-instruct.

* Fix #283 add disclaimers and bump version

* Fix up the test vector database

* Add reranker search but the test set is still incomplete.

This issue will be worked on and expanded on with #290