Speed up engine running and tests
Fix #259, working on #260. The last tests runs are still quite long, I am not sure why as the webscraping does not take long on my machine.

* Change it so the scraper only need a line in report_titles_df or report pdf downloaded.
Speeds up engine running to be basically instant.

* Fix up tests so that it has refresh equal true and will download the reports.

* Speed up tests by add in most of previous atsb safety issues.

Working on #260

* Update number of files expected in downloaded engine output

* fix #259

This only has one for TAIC and ATSB. Because TSB only takes a few seconds to scrape.

* Fix up errors